backend/requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# AI & ML
anthropic>=0.7.8
openai>=1.3.5
aiohttp>=3.9.1
numpy>=1.24.3
networkx>=3.2.1

# Audio Processing (optional - may need platform-specific installation)
# soundfile>=0.12.1
# librosa>=0.10.1
# pyaudio>=0.2.11

# Data & Storage
pydantic>=2.5.0
python-dotenv>=1.0.0
redis>=5.0.1
sqlalchemy>=2.0.23

# FetchAI uAgents
# uagents>=0.12.0  # Comment out for now as it may have compatibility issues

# Utilities
python-jose[cryptography]>=3.3.0
Pillow>=10.1.0
# opencv-python>=4.8.1.78  # Comment out for now
base64io>=1.0.3

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0